Remove verbose debugging from _obj_{un,}lock and silently discard requests if the backend is unable to acquire the lock in time
The locking code per se seems to be working correctly, but under certain conditions the frontend is simply sending too many requests concerning a single object at once. In the future it'd probably be best to rate-limit the frontend (e.g. only send the AJAX request 1 second after the most recent change to the object has been made), but for now let's dial down the debugging and just discard the requests.